Refactor add item to be controlled by internal state
