+++ src/components/Analysis.js
@@ -1,7 +1,11 @@
 import React from 'react';
+import { Outlet } from 'react-router-dom';
 
 const Analysis = () => (
-  <div>Analysis</div>
+  <div className="grid md:grid-cols-routes bg-[#fff] h-[100vh]">
+    Analysis
+    <Outlet />
+  </div>
 );
 
 export default Analysis;
