+++ src/components/History.js
@@ -1,7 +1,11 @@
 import React from 'react';
+import { Outlet } from 'react-router-dom';
 
 const History = () => (
-  <div>History</div>
+  <div className="grid md:grid-cols-routes bg-[#fff] h-[100vh]">
+    History
+    <Outlet />
+  </div>
 );
 
 export default History;
