+++ src/components/List/Header.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { BiSearchAlt2 } from 'react-icons/bi';
 
 const Header = () => (
-  <header className="flex items-center">
+  <header className="flex items-center border">
     <span className="text-[26px] font-semibold grow">
       <span className="text-orange">Shoppingify </span>
       allows you take your
