+++ tailwind.config.js
@@ -23,7 +23,8 @@ module.exports = {
     },
     extend: {
       gridTemplateColumns: {
-        desktop: '61.58px 1fr 24.3rem',
+        desktop: '61.58px 1fr',
+        routes: '1fr 24.3rem',
       },
     },
   },
